Extract RefreshCollectionsWorker; move some HomeSetDao calls to DavHomeSetRepository
