+++ app/src/main/kotlin/at/bitfire/davdroid/db/HomeSetDao.kt
@@ -36,20 +36,6 @@ interface HomeSetDao {
     @Update
     fun update(homeset: HomeSet)
 
-    /**
-     * Tries to insert new row, but updates existing row if already present.
-     * This method preserves the primary key, as opposed to using "@Insert(onConflict = OnConflictStrategy.REPLACE)"
-     * which will create a new row with incremented ID and thus breaks entity relationships!
-     *
-     * @return ID of the row, that has been inserted or updated. -1 If the insert fails due to other reasons.
-     */
-    @Transaction
-    fun insertOrUpdateByUrl(homeset: HomeSet): Long =
-        getByUrl(homeset.serviceId, homeset.url.toString())?.let { existingHomeset ->
-            update(homeset.copy(id = existingHomeset.id))
-            existingHomeset.id
-        } ?: insert(homeset)
-
     @Delete
     fun delete(homeset: HomeSet)
