+++ app/src/main/kotlin/at/bitfire/davdroid/servicedetection/RefreshCollectionsWorker.kt
@@ -21,57 +21,36 @@ import androidx.work.OutOfQuotaPolicy
 import androidx.work.WorkInfo
 import androidx.work.WorkManager
 import androidx.work.WorkerParameters
-import at.bitfire.dav4jvm.DavResource
-import at.bitfire.dav4jvm.MultiResponseCallback
-import at.bitfire.dav4jvm.Property
-import at.bitfire.dav4jvm.Response
-import at.bitfire.dav4jvm.UrlUtils
-import at.bitfire.dav4jvm.exception.HttpException
 import at.bitfire.dav4jvm.exception.UnauthorizedException
 import at.bitfire.dav4jvm.property.caldav.CalendarColor
 import at.bitfire.dav4jvm.property.caldav.CalendarDescription
-import at.bitfire.dav4jvm.property.caldav.CalendarHomeSet
-import at.bitfire.dav4jvm.property.caldav.CalendarProxyReadFor
-import at.bitfire.dav4jvm.property.caldav.CalendarProxyWriteFor
 import at.bitfire.dav4jvm.property.caldav.Source
 import at.bitfire.dav4jvm.property.caldav.SupportedCalendarComponentSet
 import at.bitfire.dav4jvm.property.carddav.AddressbookDescription
-import at.bitfire.dav4jvm.property.carddav.AddressbookHomeSet
 import at.bitfire.dav4jvm.property.carddav.SupportedAddressData
 import at.bitfire.dav4jvm.property.push.PushTransports
 import at.bitfire.dav4jvm.property.push.Topic
 import at.bitfire.dav4jvm.property.webdav.CurrentUserPrivilegeSet
 import at.bitfire.dav4jvm.property.webdav.DisplayName
-import at.bitfire.dav4jvm.property.webdav.GroupMembership
-import at.bitfire.dav4jvm.property.webdav.HrefListProperty
 import at.bitfire.dav4jvm.property.webdav.Owner
 import at.bitfire.dav4jvm.property.webdav.ResourceType
 import at.bitfire.davdroid.InvalidAccountException
 import at.bitfire.davdroid.R
 import at.bitfire.davdroid.db.AppDatabase
-import at.bitfire.davdroid.db.Collection
-import at.bitfire.davdroid.db.HomeSet
-import at.bitfire.davdroid.db.Principal
-import at.bitfire.davdroid.db.Service
 import at.bitfire.davdroid.log.Logger
 import at.bitfire.davdroid.network.HttpClient
 import at.bitfire.davdroid.servicedetection.RefreshCollectionsWorker.Companion.ARG_SERVICE_ID
 import at.bitfire.davdroid.settings.AccountSettings
-import at.bitfire.davdroid.settings.Settings
 import at.bitfire.davdroid.settings.SettingsManager
 import at.bitfire.davdroid.ui.DebugInfoActivity
 import at.bitfire.davdroid.ui.NotificationUtils
 import at.bitfire.davdroid.ui.NotificationUtils.notifyIfPossible
 import at.bitfire.davdroid.ui.account.AccountSettingsActivity
-import at.bitfire.davdroid.util.DavUtils.parent
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedInject
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.runInterruptible
-import okhttp3.HttpUrl
-import okhttp3.OkHttpClient
 import java.util.logging.Level
-import kotlin.collections.*
 
 /**
  * Refreshes list of home sets and their respective collections of a service type (CardDAV or CalDAV).
@@ -96,8 +75,9 @@ import kotlin.collections.*
 class RefreshCollectionsWorker @AssistedInject constructor(
     @Assisted appContext: Context,
     @Assisted workerParams: WorkerParameters,
-    var db: AppDatabase,
-    var settings: SettingsManager
+    val db: AppDatabase,
+    val collectionListRefresherFactory: CollectionListRefresher.Factory,
+    val settings: SettingsManager
 ): CoroutineWorker(appContext, workerParams) {
 
     companion object {
@@ -205,7 +185,7 @@ class RefreshCollectionsWorker @AssistedInject constructor(
                     .setForeground(true)
                     .build().use { client ->
                         val httpClient = client.okHttpClient
-                        val refresher = Refresher(db, service, settings, httpClient)
+                        val refresher = collectionListRefresherFactory.create(service, httpClient)
 
                         // refresh home set list (from principal url)
                         service.principal?.let { principalUrl ->
@@ -282,316 +262,6 @@ class RefreshCollectionsWorker @AssistedInject constructor(
             .build()
         NotificationManagerCompat.from(applicationContext)
             .notifyIfPossible(serviceId.toString(), NotificationUtils.NOTIFY_REFRESH_COLLECTIONS, notify)
-    }
-
-    /**
-     * Contains the methods, which do the actual refreshing work. Collected here for testability
-     */
-    class Refresher(
-        val db: AppDatabase,
-        val service: Service,
-        val settings: SettingsManager,
-        val httpClient: OkHttpClient
-    ) {
-
-        val alreadyQueried = mutableSetOf<HttpUrl>()
-
-        /**
-         * Starting at current-user-principal URL, tries to recursively find and save all user relevant home sets.
-         *
-         *
-         * @param principalUrl  URL of principal to query (user-provided principal or current-user-principal)
-         * @param level         Current recursion level (limited to 0, 1 or 2):
-         *
-         * - 0: We assume found home sets belong to the current-user-principal
-         * - 1 or 2: We assume found home sets don't directly belong to the current-user-principal
-         *
-         * @throws java.io.IOException
-         * @throws HttpException
-         * @throws at.bitfire.dav4jvm.exception.DavException
-         */
-        internal fun discoverHomesets(principalUrl: HttpUrl, level: Int = 0) {
-            Logger.log.fine("Discovering homesets of $principalUrl")
-            val relatedResources = mutableSetOf<HttpUrl>()
-
-            // Define homeset class and properties to look for
-            val homeSetClass: Class<out HrefListProperty>
-            val properties: Array<Property.Name>
-            when (service.type) {
-                Service.TYPE_CARDDAV -> {
-                    homeSetClass = AddressbookHomeSet::class.java
-                    properties = arrayOf(DisplayName.NAME, AddressbookHomeSet.NAME, GroupMembership.NAME, ResourceType.NAME)
-                }
-                Service.TYPE_CALDAV -> {
-                    homeSetClass = CalendarHomeSet::class.java
-                    properties = arrayOf(DisplayName.NAME, CalendarHomeSet.NAME, CalendarProxyReadFor.NAME, CalendarProxyWriteFor.NAME, GroupMembership.NAME, ResourceType.NAME)
-                }
-                else -> throw IllegalArgumentException()
-            }
-
-            // Query the URL
-            val principal = DavResource(httpClient, principalUrl)
-            val personal = level == 0
-            try {
-                principal.propfind(0, *properties) { davResponse, _ ->
-                    alreadyQueried += davResponse.href
-
-                    // If response holds home sets, save them
-                    davResponse[homeSetClass]?.let { homeSets ->
-                        for (homeSetHref in homeSets.hrefs)
-                            principal.location.resolve(homeSetHref)?.let { homesetUrl ->
-                                val resolvedHomeSetUrl = UrlUtils.withTrailingSlash(homesetUrl)
-                                // Homeset is considered personal if this is the outer recursion call,
-                                // This is because we assume the first call to query the current-user-principal
-                                // Note: This is not be be confused with the DAV:owner attribute. Home sets can be owned by
-                                // other principals and still be considered "personal" (belonging to the current-user-principal).
-                                db.homeSetDao().insertOrUpdateByUrl(
-                                    HomeSet(0, service.id, personal, resolvedHomeSetUrl)
-                                )
-                            }
-                    }
-
-                    // Add related principals to be queried afterwards
-                    if (personal) {
-                        val relatedResourcesTypes = listOf(
-                            // current resource is a read/write-proxy for other principals
-                            CalendarProxyReadFor::class.java,
-                            CalendarProxyWriteFor::class.java,
-                            // current resource is a member of a group (principal that can also have proxies)
-                            GroupMembership::class.java)
-                        for (type in relatedResourcesTypes)
-                            davResponse[type]?.let {
-                                for (href in it.hrefs)
-                                    principal.location.resolve(href)?.let { url ->
-                                        relatedResources += url
-                                    }
-                            }
-                    }
-
-                    // If current resource is a calendar-proxy-read/write, it's likely that its parent is a principal, too.
-                    davResponse[ResourceType::class.java]?.let { resourceType ->
-                        val proxyProperties = arrayOf(
-                            ResourceType.CALENDAR_PROXY_READ,
-                            ResourceType.CALENDAR_PROXY_WRITE,
-                        )
-                        if (proxyProperties.any { resourceType.types.contains(it) })
-                            relatedResources += davResponse.href.parent()
-                    }
-                }
-            } catch (e: HttpException) {
-                if (e.code/100 == 4)
-                    Logger.log.log(Level.INFO, "Ignoring Client Error 4xx while looking for ${service.type} home sets", e)
-                else
-                    throw e
-            }
-
-            // query related resources
-            if (level <= 1)
-                for (resource in relatedResources)
-                    if (alreadyQueried.contains(resource))
-                        Logger.log.warning("$resource already queried, skipping")
-                    else
-                        discoverHomesets(resource, level + 1)
-        }
-
-        /**
-         * Refreshes homesets and their collections.
-         *
-         * Each stored homeset URL is queried (propfind) and it's collections ([MultiResponseCallback]) either saved, updated
-         * or marked as homeless - in case a collection was removed from its homeset.
-         *
-         * If a homeset URL in fact points to a collection directly, the collection will be saved with this URL,
-         * and a null value for it's homeset. Refreshing of collections without homesets is then handled by [refreshHomelessCollections].
-         */
-        internal fun refreshHomesetsAndTheirCollections() {
-            val homesets = db.homeSetDao().getByService(service.id).associateBy { it.url }.toMutableMap()
-            for((homeSetUrl, localHomeset) in homesets) {
-                Logger.log.fine("Listing home set $homeSetUrl")
-
-                // To find removed collections in this homeset: create a queue from existing collections and remove every collection that
-                // is successfully rediscovered. If there are collections left, after processing is done, these are marked homeless.
-                val localHomesetCollections = db.collectionDao()
-                    .getByServiceAndHomeset(service.id, localHomeset.id)
-                    .associateBy { it.url }
-                    .toMutableMap()
-
-                try {
-                    DavResource(httpClient, homeSetUrl).propfind(1, *DAV_COLLECTION_PROPERTIES) { response, relation ->
-                        // Note: This callback may be called multiple times ([MultiResponseCallback])
-                        if (!response.isSuccess())
-                            return@propfind
-
-                        if (relation == Response.HrefRelation.SELF) {
-                            // this response is about the homeset itself
-                            localHomeset.displayName = response[DisplayName::class.java]?.displayName
-                            localHomeset.privBind = response[CurrentUserPrivilegeSet::class.java]?.mayBind ?: true
-                            db.homeSetDao().insertOrUpdateByUrl(localHomeset)
-                        }
-
-                        // in any case, check whether the response is about a usable collection
-                        val collection = Collection.fromDavResponse(response) ?: return@propfind
-
-                        collection.serviceId = service.id
-                        collection.homeSetId = localHomeset.id
-                        collection.sync = shouldPreselect(collection, homesets.values)
-
-                        // .. and save the principal url (collection owner)
-                        response[Owner::class.java]?.href
-                            ?.let { response.href.resolve(it) }
-                            ?.let { principalUrl ->
-                                val principal = Principal.fromServiceAndUrl(service, principalUrl)
-                                val id = db.principalDao().insertOrUpdate(service.id, principal)
-                                collection.ownerId = id
-                            }
-
-                        Logger.log.log(Level.FINE, "Found collection", collection)
-
-                        // save or update collection if usable (ignore it otherwise)
-                        if (isUsableCollection(collection))
-                            db.collectionDao().insertOrUpdateByUrlAndRememberFlags(collection)
-
-                        // Remove this collection from queue - because it was found in the home set
-                        localHomesetCollections.remove(collection.url)
-                    }
-                } catch (e: HttpException) {
-                    // delete home set locally if it was not accessible (40x)
-                    if (e.code in arrayOf(403, 404, 410))
-                        db.homeSetDao().delete(localHomeset)
-                }
-
-                // Mark leftover (not rediscovered) collections from queue as homeless (remove association)
-                for ((_, homelessCollection) in localHomesetCollections) {
-                    homelessCollection.homeSetId = null
-                    db.collectionDao().insertOrUpdateByUrlAndRememberFlags(homelessCollection)
-                }
-
-            }
-        }
-
-        /**
-         * Refreshes collections which don't have a homeset.
-         *
-         * It queries each stored collection with a homeSetId of "null" and either updates or deletes (if inaccessible or unusable) them.
-         */
-        internal fun refreshHomelessCollections() {
-            val homelessCollections = db.collectionDao().getByServiceAndHomeset(service.id, null).associateBy { it.url }.toMutableMap()
-            for((url, localCollection) in homelessCollections) try {
-                DavResource(httpClient, url).propfind(0, *DAV_COLLECTION_PROPERTIES) { response, _ ->
-                    if (!response.isSuccess()) {
-                        db.collectionDao().delete(localCollection)
-                        return@propfind
-                    }
-
-                    // Save or update the collection, if usable, otherwise delete it
-                    Collection.fromDavResponse(response)?.let { collection ->
-                        if (!isUsableCollection(collection))
-                            return@let
-                        collection.serviceId = localCollection.serviceId       // use same service ID as previous entry
-
-                        // .. and save the principal url (collection owner)
-                        response[Owner::class.java]?.href
-                            ?.let { response.href.resolve(it) }
-                            ?.let { principalUrl ->
-                                val principal = Principal.fromServiceAndUrl(service, principalUrl)
-                                val principalId = db.principalDao().insertOrUpdate(service.id, principal)
-                                collection.ownerId = principalId
-                            }
-
-                        db.collectionDao().insertOrUpdateByUrlAndRememberFlags(collection)
-                    } ?: db.collectionDao().delete(localCollection)
-                }
-            } catch (e: HttpException) {
-                // delete collection locally if it was not accessible (40x)
-                if (e.code in arrayOf(403, 404, 410))
-                    db.collectionDao().delete(localCollection)
-                else
-                    throw e
-            }
-
-        }
-
-        /**
-         * Refreshes the principals (get their current display names).
-         * Also removes principals which do not own any collections anymore.
-         */
-        internal fun refreshPrincipals() {
-            // Refresh principals (collection owner urls)
-            val principals = db.principalDao().getByService(service.id)
-            for (oldPrincipal in principals) {
-                val principalUrl = oldPrincipal.url
-                Logger.log.fine("Querying principal $principalUrl")
-                try {
-                    DavResource(httpClient, principalUrl).propfind(0, *DAV_PRINCIPAL_PROPERTIES) { response, _ ->
-                        if (!response.isSuccess())
-                            return@propfind
-                        Principal.fromDavResponse(service.id, response)?.let { principal ->
-                            Logger.log.fine("Got principal: $principal")
-                            db.principalDao().insertOrUpdate(service.id, principal)
-                        }
-                    }
-                } catch (e: HttpException) {
-                    Logger.log.info("Principal update failed with response code ${e.code}. principalUrl=$principalUrl")
-                }
-            }
-
-            // Delete principals which don't own any collections
-            db.principalDao().getAllWithoutCollections().forEach {principal ->
-                db.principalDao().delete(principal)
-            }
-        }
-
-        /**
-         * Finds out whether given collection is usable, by checking that either
-         *  - CalDAV/CardDAV: service and collection type match, or
-         *  - WebCal: subscription source URL is not empty
-         */
-        private fun isUsableCollection(collection: Collection) =
-            (service.type == Service.TYPE_CARDDAV && collection.type == Collection.TYPE_ADDRESSBOOK) ||
-                    (service.type == Service.TYPE_CALDAV && arrayOf(Collection.TYPE_CALENDAR, Collection.TYPE_WEBCAL).contains(collection.type)) ||
-                    (collection.type == Collection.TYPE_WEBCAL && collection.source != null)
-
-        /**
-         * Whether to preselect the given collection for synchronisation, according to the
-         * settings [Settings.PRESELECT_COLLECTIONS] (see there for allowed values) and
-         * [Settings.PRESELECT_COLLECTIONS_EXCLUDED].
-         *
-         * A collection is considered _personal_ if it is found in one of the current-user-principal's home-sets.
-         *
-         * Before a collection is pre-selected, we check whether its URL matches the regexp in
-         * [Settings.PRESELECT_COLLECTIONS_EXCLUDED], in which case *false* is returned.
-         *
-         * @param collection the collection to check
-         * @param homesets list of home-sets (to check whether collection is in a personal home-set)
-         * @return *true* if the collection should be preselected for synchronization; *false* otherwise
-         */
-        internal fun shouldPreselect(collection: Collection, homesets: Iterable<HomeSet>): Boolean {
-            val shouldPreselect = settings.getIntOrNull(Settings.PRESELECT_COLLECTIONS)
-
-            val excluded by lazy {
-                val excludedRegex = settings.getString(Settings.PRESELECT_COLLECTIONS_EXCLUDED)
-                if (!excludedRegex.isNullOrEmpty())
-                    Regex(excludedRegex).containsMatchIn(collection.url.toString())
-                else
-                    false
-            }
-
-            return when (shouldPreselect) {
-                Settings.PRESELECT_COLLECTIONS_ALL ->
-                    // preselect if collection url is not excluded
-                    !excluded
-
-                Settings.PRESELECT_COLLECTIONS_PERSONAL ->
-                    // preselect if is personal (in a personal home-set), but not excluded
-                    homesets
-                        .filter { homeset -> homeset.personal }
-                        .map { homeset -> homeset.id }
-                        .contains(collection.homeSetId)
-                        && !excluded
-
-                else -> // don't preselect
-                    false
-            }
-        }
     }
 
 }
